+++ migrations/20181005204935_drop_drivers.php
@@ -15,8 +15,8 @@ class DropDrivers extends Migration {
 
       $table->string('name');
 
-      $table->float('lat');
-      $table->float('lng');
+      $table->double('lat');
+      $table->double('lng');
     });
   }
 }
